Modularizing the census canned estimator example
