+++ census/estimator/trainer/metadata.py
@@ -1,0 +1,31 @@
+# Copyright 2016 Google Inc. All Rights Reserved. Licensed under the Apache
+# License, Version 2.0 (the "License"); you may not use this file except in
+# compliance with the License. You may obtain a copy of the License at
+# http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+# License for the specific language governing permissions and limitations under
+# the License.
+"""Defines a Wide + Deep model for classification on structured data.
+
+Tutorial on wide and deep: https://www.tensorflow.org/tutorials/wide_and_deep/
+"""
+
+
+# Define the format of your input data as present in the CSV file
+
+CSV_COLUMNS = [
+    'age', 'workclass', 'fnlwgt', 'education', 'education_num',
+    'marital_status', 'occupation', 'relationship', 'race', 'gender',
+    'capital_gain', 'capital_loss', 'hours_per_week', 'native_country',
+    'income_bracket'
+]
+
+CSV_COLUMN_DEFAULTS = [[0], [''], [0], [''], [0], [''], [''], [''], [''],
+                       [''], [0], [0], [0], [''], ['']]
+
+LABEL_COLUMN = 'income_bracket'
+
+LABELS = [' <=50K', ' >50K']
